Move payloads to other file
